session-15/pyproject.toml
[project]
name = "session-15"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "firekind", email = "shyamant.achar@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "jsonargparse[signatures]>=4.44.0",
    "lightning>=2.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torch-lr-finder>=0.2.2",
    "torchvision>=0.24.1",
]

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"
